updated sidebar script
