+++ Assets/Scripts/OldSidebarController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fab681ce9a5e276498eb6b83ae95ccf9
+guid: 90506e6d8a14de74881650315be4c6c3
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
